// src/types/InvExtHead.ts
export type InvExtHead = {
  /**
   * @type string
   */
  CustName?: string | null
  /**
   * @type string
   */
  CustAddresLine1?: string | null
  /**
   * @type string
   */
  CustAddresLine2?: string | null
  /**
   * @type string
   */
  CustCity?: string | null
  /**
   * @type string
   */
  CustLinePH?: string | null
  /**
   * @type string
   */
  CustMobilePH?: string | null
  /**
   * @type string
   */
  CompID?: string | null
  /**
   * @type string
   */
  Language?: string | null
  /**
   * @type string
   */
  Languge?: string | null
  /**
   * @type string
   */
  Comments?: string | null
  /**
   * @type integer | undefined, int32
   */
  CoinID?: number
  /**
   * @type string
   */
  CoinIsoName?: string | null
  /**
   * @type string
   */
  Email?: string | null
  /**
   * @type boolean | undefined
   */
  SendByEmail?: boolean
  /**
   * @type boolean
   */
  ExtIsVatFree?: boolean | null
  /**
   * @type string
   */
  ManualInvoiceNumber?: string | null
  /**
   * @type string
   */
  DepartmentId?: string | null
  /**
   * @type integer | undefined
   */
  DepartmentIdShort?: number
  /**
   * @type string
   */
  IsLoadInfoFromAccountID?: string | null
  /**
   * @type string
   */
  IsAutoCreateUpdateAccount?: string | null
  /**
   * @type string
   */
  AccountForeignKey?: string | null
  /**
   * @type string
   */
  SiteUniqueId?: string | null
  /**
   * @type string
   */
  DeliveryName?: string | null
  /**
   * @type string
   */
  InvDate?: string | null
  /**
   * @type string
   */
  ValueDate?: string | null
  /**
   * @type string
   */
  ExternalId?: string | null
  /**
   * @type string
   */
  AccountID?: string | null
}